wait for websockify to start before starting qgis
Related to https://github.com/jupyterhub/jupyter-remote-desktop-proxy/issues/105.

When QGIS starts before VNC connection is established, the QGIS process doesn't get attached to the VNC session and runs in the background.

We wait for the VNC connection to be established to make sure the QGIS GUI shows up in the VNC session.